*DO NOT USE* EXAMPLE CODE
This isn't any real code, it's just a bunch of random precompiled code
when I set it up with the new labVIEW.  The new labview setup is weird
and I need some time to figure out how to make a blank, normal robot
project.  This is to show the new programmers what a github is.