Started shooter code, added global var in begin for shooter.
Global variables in begin and other parts of begin are for the most part
complete, still need to know where they are wired. Need to ask Stefan as
to how the drivers will use the arm, although might just write code for
every way, as it would probs be faster than getting info from Stefan.