Variety of works
Silas did some stuff on the winches.

Two problems remaining with autonomous:
The first is that when I use while loops, I haven't gotten it to see the
flasher.
The second problem is that when I use structures, it shoots immediately.

I know what I have to do to fix it.