Tried fixing some logic
I tried making the logic a bit better on the autonomous, but I don't
know if it's working.  If you could test it at least once and let me
know what happens, I'd appreciate it, but please don't mess with it too
much.

Last thing, Are we even sure that the flasher works with labview that
well?  Because these tests are indicating the opposite.  I think it
worked at least once but I'm starting to wonder...