Put in semi working 2 ball
Right now I have utilized the auto selection feature so we can switch
between two ball/one ball, with the currently selected option on the
driver station.

Additionally, the two ball currently on the code(not yet shipped to the
bot) has the movements all correct, although I know it needs to be
tuned.